% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{CF}
\alias{CF}
\title{CF: Crown fraction bearing fruit}
\usage{
CF(dbh, model, It = 1)
}
\arguments{
\item{dbh}{diameter at breast height}

\item{model}{list containing the best fitting model
and a vector of coefficients}

\item{It}{Iteration for numerical sensitivity calculation.
a vector of values, corresponding to parameter dbh and of the same
length, with with to interate the vital rate function. Defaults to 1,
which means no iteration will be done. The function ItMat can be
use to build iterations. see ?ItMat.}
}
\description{
CF: Crown fraction bearing fruit
}
\author{
Marco D. Visser et al.
}

