% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{constructIPM}
\alias{constructIPM}
\title{Construct IPM}
\usage{
constructIPM(seedlingclasses, treeclasses, seedlingclasseswidth,
  treeclasseswidth, sizerange, hghtdbhthreshold, dbhthreshold, f.sp, f.sts,
  f.repro, f.sdlhght, f.s2t, f.sg, f.ss, f.tg, f.ts, Imat = Imat)
}
\arguments{
\item{seedlingclasses}{vector of seedling sizes}

\item{treeclasses}{vector of treeseedling sizes}

\item{seedlingclasseswidth}{seedling grid size}

\item{treeclasseswidth}{tree grid size}

\item{sizerange}{min and max sizes}

\item{hghtdbhthreshold}{seedling to tree threshold size in height}

\item{dbhthreshold}{seedling to tree threshold size in dbh}

\item{f.sp}{seed production model parameters}

\item{f.sts}{seed 2 seedling parameters}

\item{f.repro}{size dependent reproduction model}

\item{f.sdlhght}{initial recruit height model}

\item{f.s2t}{allometric function}

\item{f.sg}{seedling growth model}

\item{f.ss}{seedling survival model}

\item{f.tg}{tree growth model}

\item{Imat}{A iteration matrix used for numerical sensitivity calculations.
See ?ItMat.}

\item{f.ss}{tree survival model}
}
\description{
Function integrates all previous functions
and construct a full IPM
}
\author{
Marco D. Visser et al.
}

