% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{dSeedlinggrowth}
\alias{dSeedlinggrowth}
\title{dSeedlinggrowth distribution of heights t + 1 as a function
of initial height}
\usage{
dSeedlinggrowth(hghtref, hghttar, model = 4, param, sigma, gridsize, L,
  It = 1, Elt = FALSE)
}
\arguments{
\item{hghtref}{seedling initial height}

\item{hghttar}{seedling target height}

\item{model}{fitted model for seedling growth and survival}

\item{gridsize}{relavent IPM gridsize}

\item{L}{lower limit of the IPM. If NULL, seedling to 
tree transition are calculated and gridsize must be of 
the translated gridsizes according to the height transition
model!}

\item{It}{Iteration for numerical sensitivity calculation.
A vector of values, corresponding to parameter dbh/hght and of the same
length, with with to interate the vital rate function. Defaults to 1,
which means no iteration will be done. The function ItMat can be
use to build iterations. see ?ItMat.}

\item{Elt}{only return the parameter relavent for elasticity}
}
\description{
dSeedlinggrowth distribution of heights t + 1 as a function
of initial height
}
\author{
Marco D. Visser et al.
}

