% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{fecundity}
\alias{fecundity}
\title{Fecundity function}
\usage{
fecundity(dbh, S, R, F = 1, It = 1)
}
\arguments{
\item{dbh}{diameter at breast height}

\item{S}{Seed production per year per unit reproductive basal area}

\item{R}{Reproductive probabilty (of tree of size dbh)}

\item{F}{Fraction of crown bearing fruit}

\item{It}{Iteration for numerical sensitivity calculation.
a vector of values, corresponding to parameter dbh and of the same
length, with with to interate the vital rate function. Defaults to 1,
which means no iteration will be done. The function ItMat can be
use to build iterations. see ?ItMat.}
}
\description{
Fecundity function
}
\author{
Marco D. Visser et al.
}

