% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{fixedpredict}
\alias{fixedpredict}
\title{LME predict function (only fixed effects)}
\usage{
fixedpredict(x, mod = NULL, liana = NULL, sp = NULL, betas = NULL,
  time = 1)
}
\arguments{
\item{x}{design matrix}

\item{mod}{fit lme4 model and normalization coef (in a list)}

\item{liana}{which liana class?}

\item{sp}{which species to use}
}
\description{
LME predict function (only fixed effects)
}

