% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{GenTime}
\alias{GenTime}
\title{Calculate Generation time 
Returns the mean time between parent birth and offspring birth}
\usage{
GenTime(GSM, FM)
}
\arguments{
\item{GSM}{A growth-survival matrix}

\item{FM}{A fertility matrix}
}
\description{
Calculate Generation time 
Returns the mean time between parent birth and offspring birth
}

