% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{ItMat}
\alias{ItMat}
\title{Iteration matrix builder}
\usage{
ItMat(vitalrate = "tree growth", delta = 1.01, size, seedlingclasses,
  treeclasses)
}
\arguments{
\item{vitalrate}{a character specifying the vital rate
that should be iterated can be c("seed production",
 "reproduction", "recruit size", "seedling growth",
 "seedling survival", "tree growth",
 "tree survival")}

\item{delta}{the iteration factor, defualts to 1.01}

\item{size}{at which size must the iteration take place?
this is expected to be an integer corresponding to
a position in seedling or treeclasses.}

\item{seedlingclasses}{the (discretized) seedling classes
which are input into the IPM function and kernel functions}

\item{treeclasses}{the (discretized0 tree classes
which are input into the IPM function and kernel functions}
}
\description{
Constructs and iteraction matrix used in numerical
sensitivity calculations. Inputs are the vital rate,
stage and size that should be iterated.
}
\author{
Marco D. Visser et al.
}

