% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{lambda}
\alias{lambda}
\title{Retrive the dominant eigenvalue
find the dominant eigenvalue and returns it
function shamelessly stolen from the PopBio package
by Stubben et al. Which you should really be using
instead of this!}
\usage{
lambda(A, ...)
}
\arguments{
\item{A}{a square projection matrix}

\item{...}{Additional parameters passed to eigen}
}
\description{
Retrive the dominant eigenvalue
find the dominant eigenvalue and returns it
function shamelessly stolen from the PopBio package
by Stubben et al. Which you should really be using
instead of this!
}

