% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{LifeExp}
\alias{LifeExp}
\title{Calculate Mean Life Expectancy
Returns the mean life expectancy for each size}
\usage{
LifeExp(GSM)
}
\arguments{
\item{GSM}{A growth-survival matrix}
}
\description{
Calculate Mean Life Expectancy
Returns the mean life expectancy for each size
}

