% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{linearityCheck}
\alias{linearityCheck}
\title{linearityCheck}
\usage{
linearityCheck(lmeMod, xname)
}
\arguments{
\item{lmeMod}{mixed model}

\item{xname}{name of explanatory variable along which linearity
needs to be checked}
}
\description{
linearityCheck
}
\author{
Marco D. Visser et al.
}

