% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{lmepredict}
\alias{lmepredict}
\title{lmepredict: function to predict vital rates
 conform to the standard approach, but with
 more control on some processes.
LME predict function}
\usage{
lmepredict(x, mod = NULL, liana = NULL, sp = NULL, betas = NULL,
  time = 1, onlyfixedeffect = FALSE, switchoffliana = NULL,
  returngrowth = FALSE)
}
\arguments{
\item{x}{design matrix}

\item{mod}{fit lme4 model and normalization coef (in a list)}

\item{liana}{which liana class?}

\item{sp}{which species to use}

\item{time}{time correction for binomial models?}

\item{switchoffliana}{numeric: which liana cat (1:4) to switch off?}

\item{returngrowth}{return growth instead of size next?}

\item{beta}{used to provide fit with legacy code (mod = beta)}

\item{onlyfixedeffects}{logical: only used the fixed effects in predictions}
}
\description{
lmepredict: function to predict vital rates
 conform to the standard approach, but with
 more control on some processes.
LME predict function
}

