% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{NetR0}
\alias{NetR0}
\title{Calculate Net Reproductive Rate 
The net reproduction rate (NRR) is the average number of daughters
that would be born to a female (or a group of females) if she
passed through her lifetime conforming to the age-specific fertility}
\usage{
NetR0(GSM, FM)
}
\arguments{
\item{GSM}{A growth-survival matrix}

\item{FM}{A fertility matrix}
}
\description{
Calculate Net Reproductive Rate 
The net reproduction rate (NRR) is the average number of daughters
that would be born to a female (or a group of females) if she
passed through her lifetime conforming to the age-specific fertility
}

