% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{NumericElast}
\alias{NumericElast}
\title{Calculate  elasticty values from a set of lambdas returned by NumericIterate}
\usage{
NumericElast(lambdalist, lambda, delta = 1.01)
}
\arguments{
\item{lambdalist}{a list of numerically iterated lambdas and corresponding
vital rates returned by NumericIterate.}

\item{lambda}{the original eigen value}

\item{delta}{the iteration factor}

\item{vrlist}{list with the original size specific vital rates}
}
\description{
With the original lambda and a list of iterated lamdas, elasticities
are calculated
}

