% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{PassTime}
\alias{PassTime}
\title{Calculate mean passage time to a certain size
Returns the mean time from class to each size}
\usage{
PassTime(class, GSM)
}
\arguments{
\item{class}{the row/column number from which to calculate
passage time}

\item{GSM}{A growth-survival matrix}
}
\description{
Calculate mean passage time to a certain size
Returns the mean time from class to each size
}

