% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{precruitsize}
\alias{precruitsize}
\title{precruitsize: density distribution of recruits}
\usage{
precruitsize(hght, modelparam, It = 1, Elt = FALSE)
}
\arguments{
\item{hght}{seedling height}

\item{modelparam}{a list with the model name}

\item{It}{Iteration for numerical sensitivity calculation.
A vector of values, corresponding to parameter dbh/hght and of the same
length, with with to interate the vital rate function. Defaults to 1,
which means no iteration will be done. The function ItMat can be
use to build iterations. see ?ItMat.}

\item{Elt}{only return the parameter relavent for elasticity}
}
\description{
precruitsize: density distribution of recruits
}
\author{
Marco D. Visser et al.
}

