% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{pSeedlinggrowth}
\alias{pSeedlinggrowth}
\title{pSeedlinggrowth: cummulative distribution of growth  
transitions beyond a given height}
\usage{
pSeedlinggrowth(hghtref, hghttar, model, It = 1)
}
\arguments{
\item{hghtref}{seedling initial height}

\item{hghttar}{seedling target height}

\item{model}{fitted model for seedling growth}

\item{It}{Iteration for numerical sensitivity calculation.
A vector of values, corresponding to parameter dbh/hght and of the same
length, with with to interate the vital rate function. Defaults to 1,
which means no iteration will be done. The function ItMat can be
use to build iterations. see ?ItMat.}
}
\description{
pSeedlinggrowth: cummulative distribution of growth  
transitions beyond a given height
}
\author{
Marco D. Visser et al.
}

