% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{sdl2splg}
\alias{sdl2splg}
\title{sdl2splg: Seedling to sapling transition.
height to dbh translation}
\usage{
sdl2splg(hght, dbh = NULL, model, sp)
}
\arguments{
\item{hght}{hght to dbh}

\item{dbh}{dbh to height, ignored if NULL}

\item{model}{fitted parameters model for tree-height relations}

\item{sp}{species 6 letter code in capitals}
}
\description{
sdl2splg: Seedling to sapling transition.
height to dbh translation
}
\author{
Marco D. Visser et al.
}

