% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{sdlGSkernel}
\alias{sdlGSkernel}
\title{Sdlgskernel: Seedling growth and survival transitions}
\usage{
sdlGSkernel(hghtref, hghttar, models = NULL, gridsize = 10, L = 0.01,
  sImat)
}
\arguments{
\item{hghtref}{seedling initial hght,reference height}

\item{hghttar}{target height}

\item{models}{list of models for seedling growth and survival}

\item{gridsize}{seedling matrix grid size}

\item{L}{lower limit of the IPM}

\item{sImat}{A seedling iteration matrix. See ?ItMat.}

\item{dbhhghtthres}{height at 1 cm dbh}
}
\description{
Function returns transition probabilities between 
hghtref and hghttar or dbh target
}
\author{
Marco D. Visser et al.
}

