% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{Seedlingsurvival}
\alias{Seedlingsurvival}
\title{Seedlingsurvival: survival of Seedlings as a function
of initial height}
\usage{
Seedlingsurvival(hght, model, gridsize, L, It = 1)
}
\arguments{
\item{hght}{Seedling initial height}

\item{model}{list containing the fitted model function
and a set of parameters for the model}

\item{gridsize}{IPM mesh size}

\item{It}{Iteration for numerical sensitivity calculation.
A vector of values, corresponding to parameter dbh/hght and of the same
length, with with to interate the vital rate function. Defaults to 1,
which means no iteration will be done. The function ItMat can be
use to build iterations. see ?ItMat.}
}
\description{
Seedlingsurvival: survival of Seedlings as a function
of initial height
}
\author{
Marco D. Visser et al.
}

