% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{SEMatrix}
\alias{SEMatrix}
\title{SEMatrix: Calculate Sensitivity or Elasticity Matrix}
\usage{
SEMatrix(IPM, type = "elasticity")
}
\arguments{
\item{IPM}{the IPM. As square matrix}

\item{type}{either elasticity or sensitivity}
}
\description{
Function returns elasticity or sensitivity
}

