% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{tI}
\alias{tI}
\title{Transpose IPM for plotting
Function flips the IPM for correct plotting}
\usage{
tI(ipm)
}
\arguments{
\item{ipm}{the square matrix IPM}
}
\description{
Transpose IPM for plotting
Function flips the IPM for correct plotting
}

