% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{treeGSkernel}
\alias{treeGSkernel}
\title{Tree growth and survival kernel}
\usage{
treeGSkernel(dbhref, dbhtar, models, gridsize, L = 10, tImat,
  U = max(treeclasses))
}
\arguments{
\item{dbhref}{tree initial dbh,reference dbh}

\item{dbhtar}{target dbh}

\item{models}{list of models for tree growth and survival}

\item{gridsize}{seedling matrix grid size}

\item{L}{lower limit of the IP}

\item{tImat}{A tree iteration matrix. See ?ItMat.}
}
\description{
Function returns transition probabilities between 
dbhref and dbhtar
}
\author{
Marco D. Visser et al.
}

