% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{Treesurvival}
\alias{Treesurvival}
\title{Treesurvival: survival of trees as a function
of initial dbh}
\usage{
Treesurvival(dbh, model, gridsize, It = 1)
}
\arguments{
\item{dbh}{tree initial dbh}

\item{model}{fitted survival model
with first number the best model
from TSmodelnames}

\item{gridsize}{IPM mesh size for dbh}

\item{It}{Iteration for numerical sensitivity calculation.
A vector of values, corresponding to parameter dbh/hght and of the same
length, with with to interate the vital rate function. Defaults to 1,
which means no iteration will be done. The function ItMat can be
use to build iterations. see ?ItMat.}
}
\description{
Treesurvival: survival of trees as a function
of initial dbh
}
\author{
Marco D. Visser et al.
}

