% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools2sIPM.R
\name{treeTranskernel}
\alias{treeTranskernel}
\title{treeTranskernel: Seedling to tree transition matrix}
\usage{
treeTranskernel(hghtref, dbhtar, models = NULL, gridsize.tree, gridsize.sdl,
  sp = NULL, sImat)
}
\arguments{
\item{hghtref}{seedling initial hght,reference height}

\item{dbhtar}{target dbh}

\item{models}{list of models for dbh height translation, 
seedling growth and survival}

\item{sp}{6 letter species code}

\item{sImat}{A seedling iteration matrix. See ?ItMat.}

\item{dbhhghtthres}{height at 1 cm dbh}

\item{gridsize}{seedling matrix grid size (for survival)}
}
\description{
Function returns transition probabilities between 
a referebce height (hghtref) and dbh target (dbhtar)
}
\author{
Marco D. Visser et al.
}

